+++ botlink/src/lib.rs
@@ -1,15 +1,6 @@
 #[macro_use]
 extern crate lazy_static;
-extern crate base64;
-extern crate bincode;
-extern crate crossbeam;
-extern crate crossbeam_channel;
-extern crate dotenv;
-extern crate env_logger;
-extern crate log;
-extern crate micro_model_bot;
-extern crate micro_model_moves;
-extern crate redis_streams;
+
 pub extern crate tokio;
 extern crate tokio_tungstenite;
 extern crate uuid;