+++ botlink/src/registry.rs
@@ -1,23 +1,29 @@
 use crate::repo::*;
+use crate::stream::xack::XAck;
 use crate::stream::xadd::*;
-use crate::stream::xread::{RedisXReader, XReader};
+use crate::stream::xread::XReader;
 use crossbeam_channel::{unbounded, Receiver, Sender};
 use micro_model_bot::{ComputeMove, MoveComputed};
-use redis_conn_pool;
-use redis_conn_pool::RedisHostUrl;
+use redis::Client;
 use std::sync::Arc;
 
 pub struct Components {
     pub ab_repo: Box<dyn AttachedBotsRepo>,
-    pub entry_id_repo: Box<dyn EntryIdRepo>,
     pub board_size_repo: Arc<dyn BoardSizeRepo>,
     pub xreader: Box<dyn XReader>,
     pub xadder: Arc<dyn XAdder>,
+    pub xack: Arc<dyn XAck>,
     pub compute_move_in: Sender<ComputeMove>,
     pub compute_move_out: Receiver<ComputeMove>,
     pub move_computed_in: Sender<MoveComputed>,
     pub move_computed_out: Receiver<MoveComputed>,
 }
+
+const REDIS_URL: &str = "redis://redis/";
+pub fn create_redis_client() -> Arc<redis::Client> {
+    Arc::new(Client::open(REDIS_URL).expect("redis client"))
+}
+
 impl Default for Components {
     fn default() -> Self {
         let (compute_move_in, compute_move_out): (Sender<ComputeMove>, Receiver<ComputeMove>) =
@@ -26,13 +32,13 @@ impl Default for Components {
         let (move_computed_in, move_computed_out): (Sender<MoveComputed>, Receiver<MoveComputed>) =
             unbounded();
 
-        let pool = Arc::new(redis_conn_pool::create(RedisHostUrl::default()));
+        let client = create_redis_client();
         Components {
-            ab_repo: Box::new(pool.clone()),
-            entry_id_repo: Box::new(pool.clone()),
-            board_size_repo: Arc::new(pool.clone()),
-            xreader: Box::new(RedisXReader { pool: pool.clone() }),
-            xadder: Arc::new(RedisXAdder { pool }),
+            ab_repo: Box::new(client.clone()),
+            board_size_repo: Arc::new(client.clone()),
+            xreader: Box::new(client.clone()),
+            xadder: Arc::new(client.clone()),
+            xack: Arc::new(client),
             compute_move_in,
             compute_move_out,
             move_computed_in,